* commit 'dd384859b6686909e783c8b714accdea6a962c04':
  Fix NativeBreakIterator_cloneImpl return type.

+++ luni/src/main/native/libcore_icu_NativeBreakIterator.cpp
@@ -119,7 +119,7 @@ class BreakIteratorAccessor {
   } \
   return reinterpret_cast<uintptr_t>(it)
 
-static jint NativeBreakIterator_cloneImpl(JNIEnv* env, jclass, jlong address) {
+static jlong NativeBreakIterator_cloneImpl(JNIEnv* env, jclass, jlong address) {
   BreakIteratorAccessor it(env, address);
   return reinterpret_cast<uintptr_t>(it->clone());
 }
